worked on compiler warnings
git-svn-id: svn://svn.code.sf.net/p/pmii/code/trunk@5023 2dc5e0b7-7910-0410-9abd-fb16afba6634

+++ QedeqKernelBo/src/org/qedeq/kernel/bo/logic/proof/checker/ProofChecker1Impl.java
@@ -149,7 +149,7 @@ public class ProofChecker1Impl implements ProofChecker {
                 if (n != null) {
                     final ModuleContext lc = new ModuleContext(moduleContext.getModuleLocation(),
                         moduleContext.getLocationWithinModule() + ".get("
-                        + ((Integer) label2line.get(line.getLabel()))
+                        + label2line.get(line.getLabel())
                         + ").getLabel()");
                     setLocationWithinModule(context + ".get("  + i + ").getLabel()");
                     handleProofCheckException(
